# swagger.yaml
swagger: "2.0"
info:
  description: "This is a business web application."
  version: "1.0.0"
  title: "WeConnect Web Application"
  termsOfService: "http://swagger.io/terms/"
  contact:
    email: "augustineezinwa@gmail.com"
  license:
    name: "Apache 2.0"
    url: "http://www.apache.org/licenses/LICENSE-2.0.html"
host: "we--connect.herokuapp.com"
basePath: "/api/v1"
...
schemes:
- "https"
paths:
...
  /businesses/{businessId}:
    get:
      tags:
      - "Businesses"
      summary: "Get a business with an id parameter"
      produces:
      - "application/json"
      parameters:
      - name: "businessId"
        in: "path"
        description: "Id to fetch a business"
        required: true
        type: "integer"
      responses:
        200:
          description: "business search was successful"
          schema:
            $ref: "#/definitions/BusinessFoundModel"
        404:
          description: "Business not found"
          
    put:
      tags:
      - "Businesses"
      summary: "Update a business profile"
      description: "This can only be done by the logged in user."
      operationId: "updateBusiness"
      produces:
      - "application/json"
      parameters:
      - name: "businessId"
        in: "path"
        description: "business that needs to be updated"
        required: true
      - name: "body"
        in: "body"
        description: "business Object to database"
        required: true
        schema:
          $ref: "#/definitions/RegisterBusinessModel"
      responses:
        200:
          description: "Business Update Successful"
        404:
          description: "Business with id does not exist"
    delete:
      tags:
      - "Businesses"
      summary: "Delete Business"
      description: "This can only be done by the logged in user."
      operationId: "deleteBusiness"
      produces:
      - "application/json"
      parameters:
      - name: "businessId"
        in: "path"
        description: "The id of the business that needs to be deleted"
        required: true
        type: "integer"
      responses:
        206:
          description: "No Content"
        404:
          description: "Business with id does not exist"
...
definitions:
...
  RegisterBusinessModel:
    type: "object"
    required: ["businessName", "businessAddress","businessDescripton","phoneNumber","email", "location", "category"]
    properties:
      businessName:
        type: "string"
        uniqueItems: true
      businessAddress:
        type: "string"
        uniqueItems: true
      businessDescription:
        type: "string"
        uniqueItems: true
      email:
        type: "string"
        uniqueItems: true
      phoneNumber:
        type: "string"
        uniqueItems: true
      location:
        type: "string"
        uniqueItems: false
      userId:
        type: "string"
        uniqueItems: true
      category:
        type: "string"
        uniqueItems: true
  BusinessFoundModel:
    type: "object"
    required: ["business"]
    properties:
      message:
       type: "string"
      business:
        type: "object"
...